Refactor 4
