--- Database/makeFile
@@ -1,15 +1,0 @@
-DB=database.bin
-SOURCES_NP=$(wildcard ../Api/namedPipe/*.c)
-SOURCES_ST=$(wildcard ../Api/socket/*.c)
-SOURCES_DB=$(wildcard ./*.c)
-
-allST: $(SOURCES_ST) $(SOURCES_DB)
-	gcc $(SOURCES_ST) $(SOURCES_DB) -o $(DB) -lsqlite3 -Wall
-
-allNP: $(SOURCES_NP) $(SOURCES_DB)
-	gcc $(SOURCES_NP) $(SOURCES_DB) -o $(DB) -lsqlite3 -Wall
-
-clean:
-	rm -rf  *.o *.bin 
-
-.PHONY: clean
