+++ Server/Makefile
@@ -3,7 +3,7 @@ SOURCES_API=$(wildcard ../Api/*.c)
 SOURCES_NP=$(wildcard ../Api/namedPipe/*.c)
 SOURCES_ST=$(wildcard ../Api/socket/*.c)
 SOURCES_SERVER=$(wildcard ./*.c)
-SOURCES_DB=$(wildcard ../Database/databaseapi.c)
+SOURCES_DB=$(wildcard ../Database/*.c)
 
 allST: $(SOURCES_ST) $(SOURCES_SERVER) $(SOURCES_API)
 	gcc $(SOURCES_ST) $(SOURCES_SERVER) $(SOURCES_API) -o $(SERVER) -Wall
